tools/requirements.txt
# ML + tagging model
torch>=2.0.0
timm>=0.9.0
huggingface-hub>=0.13.3

# Data handling
numpy>=1.23
pandas>=1.5

# Image processing
pillow>=9.0.0
pyvips>=2.2

# Parsing / CLI args
simple-parsing>=0.1.3

# API and HTTP
requests>=2.28

# SQLite + progress
tqdm>=4.64.0

# JSON (optional speedup)
orjson>=3.8.0 ; python_version >= "3.8"

# PostgreSQL
psycopg2-binary>=2.9.5

tk
